merge branch "selena" onto "prototype" - all good

+++ LibrarySystem-Backend/src/main/java/ca/mcgill/ecse321/librarysystem/service/RoomService.java
@@ -21,12 +21,36 @@ public class RoomService {
 	@Autowired
 	HeadLibrarianRepository headLibrarianRepository;
 	
+	
 	// creates room, returns it so we know it's not null 
 	@Transactional 
 	public Room createRoom(
+			String currentUserId,
 			String roomNum,
-		    Integer capacity) 
+		    Integer capacity)  throws Exception
 	{
+		// check capacity 
+		if (capacity<1) throw new IllegalArgumentException("Capacity must be at least 1");
+		
+		// check permission: only librarians have permission to create the rooms
+		HeadLibrarian currentHeadLibrarian = headLibrarianRepository.findUserByIdNum(currentUserId);
+		Librarian currentLibrarian = librarianRepository.findUserByIdNum(currentUserId);
+	    if (
+	      (currentHeadLibrarian != null && currentHeadLibrarian.getIsLoggedIn() ) || ( currentLibrarian != null && currentLibrarian.getIsLoggedIn() )
+	    ) {
+	    
+	    }else {
+	    	 throw new IllegalArgumentException(
+	    		        "You do not have permission to create a room"
+	    		      );
+	    }
+	    
+	    // check if newRoomNum is available
+	    for (Room room: toList(roomRepository.findAll()) )  {
+	    	if ( room.getRoomNum().equalsIgnoreCase(roomNum) ) throw new IllegalArgumentException("The room number already exists");
+	    }
+	    
+	    
 		Room room = new Room();
 	    room.setRoomNum(roomNum);
 	    room.setCapacity(capacity);
@@ -48,32 +72,61 @@ public class RoomService {
 	}
 
 	
-	// TODO add method for modify rooms
-	// only let head librarian modify rooms
-	
-	// TODO add method for delete rooms
+	// method for modify rooms
+	// only head librarian have permission to modify the rooms, they can modify the room number and the capacity, but the room numbers must be unique
 	@Transactional 
-	public Room deleteRoom (String currentUserId, String roomNum) 
+	public Room updateRoom (String currentUserId, String oldRoomNum, String newRoomNum, int newCapacity) throws Exception
 	{
-		Librarian currentLibrarian = librarianRepository.findUserByIdNum(
-			      currentUserId
-			    );
-			    HeadLibrarian currentHeadLibrarian = headLibrarianRepository.findUserByIdNum(
-			      currentUserId
-			    );
-			    if (
-			      currentLibrarian == null ||
-			      !currentLibrarian.getIsLoggedIn() &&
-			      (currentHeadLibrarian == null || !currentHeadLibrarian.getIsLoggedIn())
-			    ) {
-			      throw new IllegalArgumentException(
-			        "You do not have permission to create an item reservation"
-			      );
-			    }
+		// check capacity 
+		if (newCapacity<1) throw new IllegalArgumentException("Capacity must be at least 1");
+		
+		// check permission: only librarians have permission to update the rooms
+		HeadLibrarian currentHeadLibrarian = headLibrarianRepository.findUserByIdNum(currentUserId);
+		Librarian currentLibrarian = librarianRepository.findUserByIdNum(currentUserId);
+	    if (
+	      (currentHeadLibrarian != null && currentHeadLibrarian.getIsLoggedIn() ) || ( currentLibrarian != null && currentLibrarian.getIsLoggedIn() )
+	    ) {
+	    
+	    }else {
+	    	 throw new IllegalArgumentException(
+	    		        "You do not have permission to update a room"
+	    		      );
+	    }
+		
+	    // check if newRoomNum is available
+	    for (Room room: toList(roomRepository.findAll()) )  {
+	    	if ( room.getRoomNum().equalsIgnoreCase(newRoomNum) ) throw new IllegalArgumentException("The new room number already exists");
+	    }
+	    
+	    
+		Room toUpdate = roomRepository.findRoomByRoomNum(oldRoomNum);
+		toUpdate.setCapacity(newCapacity);
+		toUpdate.setRoomNum(newRoomNum);
+		
+
+	    return toUpdate;		
+	}
+	
+	// method for delete rooms
+	@Transactional 
+	public Room deleteRoom (String currentUserId, String roomNum) throws Exception
+	{
+		// check permission: only librarians have permission to delete rooms
+		HeadLibrarian currentHeadLibrarian = headLibrarianRepository.findUserByIdNum(currentUserId);
+		Librarian currentLibrarian = librarianRepository.findUserByIdNum(currentUserId);
+	    if (
+	      (currentHeadLibrarian != null && currentHeadLibrarian.getIsLoggedIn() ) || ( currentLibrarian != null && currentLibrarian.getIsLoggedIn() )
+	    ) {
+	    
+	    }else {
+	    	 throw new IllegalArgumentException(
+	    		        "You do not have permission to delete a room"
+	    		      );
+	    }
 			    
 		Room toDelete = roomRepository.findRoomByRoomNum(roomNum);
+		if (toDelete == null) throw new IllegalArgumentException("The room number is invalid");
 		roomRepository.delete(toDelete);
-
 	    return toDelete;		
 	}
